cleanup and fixing lints, adding todo items.

+++ crabroll/src/motor.rs
@@ -51,7 +51,7 @@ pub(crate) async fn motor_task(
             }
             Command::MoveToPos(percent) => {
                 info!("moving to {}", percent);
-                let pos = (percent as u32 * stepper.travel_limit().get()) / 100 as u32;
+                let pos = (percent as u32 * stepper.travel_limit().get()) / 100_u32;
                 info!("moving to {}", pos);
                 match execute_move(&mut step_pin, &mut dir_pin, &mut stepper, pos).await {
                     Ok(_) => info!("moved to pos"),
@@ -63,7 +63,7 @@ pub(crate) async fn motor_task(
             }
         }
         CURRENT_POS.signal(if let Some(p) = stepper.pos() {
-            ((p * 100 as u32) / stepper.travel_limit())
+            ((p * 100_u32) / stepper.travel_limit())
                 .try_into()
                 .unwrap_or(100)
         } else {
